requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
tqdm==4.66.1